data_fundamentals/tp4/requirements.txt
mysql-connector-python==9.2.0
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
